+++ include/mrc_mat.h
@@ -11,7 +11,7 @@
 MRC_CLASS_DECLARE(mrc_mat, struct mrc_mat);
 
 void mrc_mat_assemble(struct mrc_mat *mat);
-void mrc_mat_add_value(struct mrc_mat *mat, int row_idx, int col_idx, float val);
+void mrc_mat_add_value(struct mrc_mat *mat, int row_idx, int col_idx, double val);
 void mrc_mat_apply(struct mrc_fld *y, struct mrc_mat *mat, struct mrc_fld *x);
 void mrc_mat_apply_in_place(struct mrc_mat *mat, struct mrc_fld *x);
